fix(netplay): the LOSSY_ADAPTER test facility had stopped compiling
`random_float()` returns `(float)rand() / RAND_MAX`. RAND_MAX (0x7fffffff)
is not representable as a float, so the implicit int->float conversion trips
-Wimplicit-const-int-float-conversion, and this project builds -Werror. The
whole facility therefore failed to compile the moment the warning set
tightened, and nobody noticed because it sits behind `#define LOSSY_ADAPTER`,
which is commented out and which no gate configures.

The consequence was worse than a dead #define: there was NO way to exercise
rollback in a local test. Loopback netplay has no loss and no latency, so a
two-instance session runs at rb_hist=0:60 forever — it proves the transport
and the handshake, and says nothing about save/restore, the desync checksum,
or prediction. The one facility built to force rollbacks could not be built.

Cast explicitly. The value rounds to 2147483648.0f either way, which is what
the ratio wants. Compiled out of every shipped and gated build, so this
changes no artifact.

With it working, a 25% send-loss soak on both peers ran 10,755 frames with
real rollbacks (rb_hist buckets 1 and 2 populated — 21 single-frame and one
two-frame rollback inside a single second), both sides still frame-locked at
the end, and ZERO desyncs. That exercises what the loopback soak could not:
thousands of save_state/load_state round trips through the sparse path, the
focused checksum recomputed and compared across every rollback, and the
bounded session_health map accepting legitimate frames throughout.

Also repoints the one enforced citation these six lines drifted
(netplay.c:2548-2552 -> :2553-2557 for the CONNECTING deadline arming),
verified by reading the new range rather than with --fix.

Gates GREEN, harness 13; arm-cross-build NOT RUN. Baselines 11/0/0.

+++ src/netplay/netplay.c
@@ -925,7 +925,12 @@ static GekkoNetAdapter* base_adapter = NULL;
 static GekkoNetAdapter lossy_adapter = { 0 };
 
 static float random_float() {
-    return (float)rand() / RAND_MAX;
+    /* RAND_MAX (0x7fffffff) is not representable as a float, so the implicit
+     * int->float conversion trips -Wimplicit-const-int-float-conversion under
+     * -Werror and this whole facility stopped compiling when the warning set
+     * tightened. Cast explicitly: the value rounds to 2147483648.0f either
+     * way, which is what the ratio wants. */
+    return (float)rand() / (float)RAND_MAX;
 }
 
 static void LossyAdapter_SendData(GekkoNetAddress* addr, const char* data, int length) {
